created a separate ppglm directory, added validation
