+++ src/nemos/pp_glm/__init__.py
@@ -1,0 +1,1 @@
+"""Poisson Process GLM module."""
